支持 在自己的项目中配置epii-server.ini

+++ install/install.php
@@ -44,6 +44,44 @@ if(!( ( stripos($ini['server']['www_dir'],"/")===0) || ( stripos($ini['server'][
     $ini['server']['www_dir'] = $base_root . DIRECTORY_SEPARATOR.$ini['server']['www_dir'];
 }
 
+if(isset($ini['server']['www_dir']) && is_dir($ini['server']['www_dir']))
+{
+    $file_arr = scandir($ini['server']['www_dir']);
+    
+    foreach($file_arr as $item){
+
+        if($item!=".." && $item !="."){
+
+            if(is_dir($tmp_dir = $ini['server']['www_dir']."/".$item)){
+                if(file_exists($tmp_config_file =$tmp_dir."/epii-server.ini" )){
+                    
+                    $this_ini = parse_ini_file($tmp_config_file,true );
+                    if($this_ini){
+                        $this_ini = array_merge(["name"=>$item,"root"=>""],$this_ini);
+                        $ini['app_dir'][$this_ini["name"]] = $tmp_dir."/".$this_ini["root"];
+                        if(isset($this_ini["php_select"])){
+                            $ini['app_php_select'][$this_ini["name"]] = $this_ini["php_select"];
+                        }
+                        if(isset($this_ini["server_name"])){
+                            $_domains= explode(" ",$this_ini["server_name"]);
+                            foreach($_domains as $value)
+                            {
+                                if($value=trim($value))
+                                {
+                                    $ini['domain_app'][$value] = $this_ini["name"];
+                                }
+                            }
+                            
+                        }
+                    }
+                }
+
+            } 
+        }
+    }
+}
+
+ 
 $find = ["domain_app", "base_root", "this_ip", "this_port",  "nginx_root", "domain_this", "domain_other", "domain_this_1", "domain_other_1", "www_dir", "nginx_cmd", "php_cmd"];
 $replace = [implode(" ", array_keys($ini['domain_app'])), $base_root, $ini['server']['this_ip'], $ini['server']['this_port'], $ini['server']['nginx_root'], $ini['server']['domain_this'], $ini['server']['domain_other'], str_replace(".", "\\.", $ini['server']['domain_this']), str_replace(".", "\\.", $ini['server']['domain_other']), isset($ini['server']['www_dir']) ? $ini['server']['www_dir'] : $base_root . DIRECTORY_SEPARATOR . "web", $ini['nginx']['cmd'], $ini['server']['php_cmd']];
 
@@ -176,6 +214,7 @@ if (!$is_win) {
     chmod($start_bat, 0777);
     chmod($nignx_config, 0777);
     chmod($epii_server_bat, 0777);
+    if(!file_exists("/usr/local/bin/epii-server"))
     system("ln -s ".$epii_server_bat." /usr/local/bin/epii-server");
 
 }
